Adding support for decimal as data type.
Conflicts:
	lib/Cake/View/Helper/FormHelper.php

+++ lib/Cake/Test/Case/Model/Datasource/Database/MysqlTest.php
@@ -541,7 +541,7 @@ class MysqlTest extends CakeTestCase {
 		$this->assertEquals($expected, $result);
 
 		$result = $this->Dbo->column('decimal(14,7) unsigned');
-		$expected = 'float';
+		$expected = 'decimal';
 		$this->assertEquals($expected, $result);
 	}
